internship_project/requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
transformers==4.34.0
torch>=2.0.0
sentencepiece==0.1.99
accelerate==0.22.0
